requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.7.5
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
